requirements.txt
streamlit
opencv-python
Pillow
numpy
easyocr